+++ test/core/dialects/snowflake_test.py
@@ -3,7 +3,10 @@
 import pytest
 
 from sqlfluff.core import Linter
-from sqlfluff.core.dialects.dialect_snowflake import UseStatementSegment
+from sqlfluff.core.dialects.dialect_snowflake import (
+    UseStatementSegment,
+    SemiStructuredAccessorSegment,
+)
 
 
 @pytest.mark.parametrize(
@@ -16,6 +19,11 @@ from sqlfluff.core.dialects.dialect_snowflake import UseStatementSegment
         (UseStatementSegment, 'USE SCHEMA "MY_DATABASE"."MY_SCHEMA";'),
         (UseStatementSegment, 'USE SCHEMA "MY_SCHEMA";'),
         (UseStatementSegment, 'USE "MY_DATABASE"."MY_SCHEMA";'),
+        # Testing https://github.com/sqlfluff/sqlfluff/issues/634
+        (
+            SemiStructuredAccessorSegment,
+            "SELECT ID :: VARCHAR as id, OBJ : userId :: VARCHAR as user_id from x",
+        ),
     ],
 )
 def test_snowflake_queries(segment_cls, raw, caplog):